[SCRUM-148]:useCanvasInteradction.ts로 이밴트 핸들러 분리
