+++ src/components/canvas/PixelCanvas.tsx
@@ -1,12 +1,12 @@
 import React, { useRef, useEffect, useCallback, useState } from 'react';
 import { useCanvasUiStore } from '../../store/canvasUiStore';
-import { shallow } from 'zustand/shallow';
 import { usePixelSocket } from '../SocketIntegration';
 import CanvasUI from './CanvasUI';
 import Preloader from '../Preloader';
 import { useCanvasStore } from '../../store/canvasStore';
 import { toast } from 'react-toastify';
 import { fetchCanvasData as fetchCanvasDataUtil } from '../../api/canvasFetch';
+import { useCanvasInteraction } from '../../hooks/useCanvasInteraction';
 
 import {
   INITIAL_POSITION,
@@ -28,28 +28,14 @@ function PixelCanvas({
 }: PixelCanvasProps) {
   const { canvas_id, setCanvasId } = useCanvasStore();
 
-  useEffect(() => {
-    if (initialCanvasId && initialCanvasId !== canvas_id) {
-      setCanvasId(initialCanvasId);
-      console.log('Canvas ID changed:', initialCanvasId);
-    }
-  }, [initialCanvasId, canvas_id, setCanvasId]);
-
   const rootRef = useRef<HTMLDivElement>(null);
   const previewCanvasRef = useRef<HTMLCanvasElement>(null);
   const renderCanvasRef = useRef<HTMLCanvasElement>(null);
   const interactionCanvasRef = useRef<HTMLCanvasElement>(null);
   const sourceCanvasRef = useRef<HTMLCanvasElement>(null!);
-
   const scaleRef = useRef<number>(1);
   const viewPosRef = useRef<{ x: number; y: number }>(INITIAL_POSITION);
-  const startPosRef = useRef<{ x: number; y: number }>(INITIAL_POSITION);
-  const isPanningRef = useRef<boolean>(false);
-
-  const pinchDistanceRef = useRef<number>(0);
-  const dragStartInfoRef = useRef<{ x: number; y: number } | null>(null);
   const DRAG_THRESHOLD = 5; // 5px 이상 움직이면 드래그로 간주
-
   const fixedPosRef = useRef<{ x: number; y: number; color: string } | null>(
     null
   );
@@ -59,18 +45,14 @@ function PixelCanvas({
     color: string;
   } | null>(null);
 
+  const imageTransparencyRef = useRef(0.5);
+
   // state를 각각 가져오도록 하여 불필요한 리렌더링을 방지합니다.
   const [canvasSize, setCanvasSize] = useState({ width: 0, height: 0 });
 
   const color = useCanvasUiStore((state) => state.color);
-  const setColor = useCanvasUiStore((state) => state.setColor);
-  const hoverPos = useCanvasUiStore((state) => state.hoverPos);
   const setHoverPos = useCanvasUiStore((state) => state.setHoverPos);
   const cooldown = useCanvasUiStore((state) => state.cooldown);
-  const setCooldown = useCanvasUiStore((state) => state.setCooldown);
-  const timeLeft = useCanvasUiStore((state) => state.timeLeft);
-  const setTimeLeft = useCanvasUiStore((state) => state.setTimeLeft);
-  const showPalette = useCanvasUiStore((state) => state.showPalette);
   const setShowPalette = useCanvasUiStore((state) => state.setShowPalette);
   const showImageControls = useCanvasUiStore(
     (state) => state.showImageControls
@@ -90,17 +72,12 @@ function PixelCanvas({
   );
   const isLoading = useCanvasUiStore((state) => state.isLoading);
   const setIsLoading = useCanvasUiStore((state) => state.setIsLoading);
-  const hasError = useCanvasUiStore((state) => state.hasError);
-  const setHasError = useCanvasUiStore((state) => state.setHasError);
   const showCanvas = useCanvasUiStore((state) => state.showCanvas);
   const setShowCanvas = useCanvasUiStore((state) => state.setShowCanvas);
   const targetPixel = useCanvasUiStore((state) => state.targetPixel);
   const setTargetPixel = useCanvasUiStore((state) => state.setTargetPixel);
 
   const startCooldown = useCanvasUiStore((state) => state.startCooldown);
-
-  const imageTransparencyRef = useRef(0.5);
-  const lastTouchPosRef = useRef<{ x: number; y: number } | null>(null);
 
   // 이미지 관련 상태 (Zustand로 이동하지 않는 부분)
   const imageCanvasRef = useRef<HTMLCanvasElement>(null);
@@ -159,7 +136,7 @@ function PixelCanvas({
       }
       return null;
     },
-    [imagePosition, imageSize, isImageFixed]
+    [imagePosition, imageSize, isImageFixed, scaleRef]
   );
 
   const draw = useCallback(() => {
@@ -370,7 +347,7 @@ function PixelCanvas({
 
       img.src = URL.createObjectURL(file);
     },
-    [canvasSize, draw]
+    [canvasSize, draw, setIsImageFixed, setShowImageControls, setShowPalette]
   );
 
   // 이미지 확대축소
@@ -404,7 +381,7 @@ function PixelCanvas({
     setIsImageFixed(true);
     setShowImageControls(false);
     toast.success('이미지가 고정되었습니다!');
-  }, []);
+  }, [setIsImageFixed, setShowImageControls]);
 
   // 이미지 취소
   const cancelImage = useCallback(() => {
@@ -413,7 +390,7 @@ function PixelCanvas({
     setIsImageFixed(false);
     toast.info('이미지가 제거되었습니다.');
     draw();
-  }, [draw]);
+  }, [draw, setIsImageFixed, setShowImageControls]);
 
   const { sendPixel } = usePixelSocket({
     sourceCanvasRef,
@@ -461,7 +438,7 @@ function PixelCanvas({
         setHoverPos(null);
       }
     },
-    [canvasSize]
+    [canvasSize, viewPosRef, scaleRef, setHoverPos, interactionCanvasRef]
   );
 
   const clearOverlay = useCallback(() => {
@@ -470,7 +447,7 @@ function PixelCanvas({
     if (!overlayCanvas) return;
     const overlayCtx = overlayCanvas.getContext('2d');
     overlayCtx?.clearRect(0, 0, overlayCanvas.width, overlayCanvas.height);
-  }, []);
+  }, [setHoverPos, interactionCanvasRef]);
 
   const resetAndCenter = useCallback(() => {
     const canvas = renderCanvasRef.current;
@@ -479,14 +456,11 @@ function PixelCanvas({
     // 화면 크기에 맞게 스케일 계산
     const viewportWidth = canvas.clientWidth;
     const viewportHeight = canvas.clientHeight;
-    const isMobile = window.innerWidth < 768;
 
     const scaleFactor = 0.7;
     const scaleX = (viewportWidth / canvasSize.width) * scaleFactor;
     const scaleY = (viewportHeight / canvasSize.height) * scaleFactor;
     scaleRef.current = Math.max(Math.min(scaleX, scaleY), MIN_SCALE);
-    scaleRef.current = Math.min(scaleRef.current, MAX_SCALE);
-    scaleRef.current = Math.min(scaleX, scaleY);
     scaleRef.current = Math.min(scaleRef.current, MAX_SCALE);
 
     // 캔버스를 화면 중앙에 배치
@@ -497,7 +471,7 @@ function PixelCanvas({
 
     draw();
     clearOverlay();
-  }, [draw, clearOverlay, canvasSize]);
+  }, [draw, clearOverlay, canvasSize, scaleRef, viewPosRef, renderCanvasRef]);
 
   const centerOnPixel = useCallback(
     (screenX: number, screenY: number) => {
@@ -549,7 +523,7 @@ function PixelCanvas({
       };
       requestAnimationFrame(animate);
     },
-    [draw, updateOverlay, canvasSize]
+    [draw, updateOverlay, canvasSize, viewPosRef, scaleRef, renderCanvasRef]
   );
 
   const zoomCanvas = useCallback(
@@ -575,7 +549,7 @@ function PixelCanvas({
       draw();
       updateOverlay(centerX, centerY);
     },
-    [draw, updateOverlay]
+    [draw, updateOverlay, viewPosRef, scaleRef, renderCanvasRef]
   );
 
   const handleZoomIn = useCallback(() => {
@@ -648,7 +622,7 @@ function PixelCanvas({
       };
       requestAnimationFrame(animate);
     },
-    [draw, canvasSize, updateOverlay]
+    [draw, canvasSize, updateOverlay, viewPosRef, scaleRef, renderCanvasRef]
   );
 
   const handleCooltime = useCallback(() => {
@@ -679,210 +653,54 @@ function PixelCanvas({
     [draw]
   );
 
-  const handleMouseDown = useCallback(
-    (e: React.MouseEvent<HTMLCanvasElement>) => {
-      const sx = e.nativeEvent.offsetX;
-      const sy = e.nativeEvent.offsetY;
-      const wx = (sx - viewPosRef.current.x) / scaleRef.current;
-      const wy = (sy - viewPosRef.current.y) / scaleRef.current;
-
-      // 이미지 모드에서 리사이즈 핸들 또는 이미지 영역 클릭 감지
-      if (
-        imageMode &&
-        !isImageFixed &&
-        imageCanvasRef.current &&
-        e.button === 0
-      ) {
-        const handle = getResizeHandle(wx, wy);
-
-        if (handle) {
-          // 리사이즈 핸들 클릭
-          setIsResizing(true);
-          setResizeHandle(handle);
-          setResizeStart({
-            x: wx,
-            y: wy,
-            width: imageSize.width,
-            height: imageSize.height,
-          });
-          return;
-        } else if (
-          wx >= imagePosition.x &&
-          wx <= imagePosition.x + imageSize.width &&
-          wy >= imagePosition.y &&
-          wy <= imagePosition.y + imageSize.height
-        ) {
-          // 이미지 드래그
-          setIsDraggingImage(true);
-          setDragStart({ x: wx - imagePosition.x, y: wy - imagePosition.y });
-          return;
-        }
-      }
-
-      if (e.button === 0) {
-        dragStartInfoRef.current = { x: sx, y: sy };
-      }
-    },
-    [
-      imageMode,
-      isImageFixed,
-      imagePosition,
-      imageSize,
-      getResizeHandle,
-      setIsResizing,
-      setResizeHandle,
-      setResizeStart,
-      setIsDraggingImage,
-      setDragStart,
-    ]
-  );
-
-  const handleMouseMove = useCallback(
-    (e: React.MouseEvent<HTMLCanvasElement>) => {
-      const { offsetX, offsetY } = e.nativeEvent;
-
-      if (dragStartInfoRef.current && !isPanningRef.current) {
-        const dx = offsetX - dragStartInfoRef.current.x;
-        const dy = offsetY - dragStartInfoRef.current.y;
-        if (Math.sqrt(dx * dx + dy * dy) > DRAG_THRESHOLD) {
-          isPanningRef.current = true;
-          startPosRef.current = {
-            // 패닝 시작 위치 설정
-            x: offsetX - viewPosRef.current.x,
-            y: offsetY - viewPosRef.current.y,
-          };
-          dragStartInfoRef.current = null; // 대기 상태 해제
-        }
-      }
-
-      // 이미지 리사이즈 중
-      if (isResizing && resizeHandle) {
-        const wx = (offsetX - viewPosRef.current.x) / scaleRef.current;
-        const wy = (offsetY - viewPosRef.current.y) / scaleRef.current;
-
-        let newWidth = imageSize.width;
-        let newHeight = imageSize.height;
-
-        if (resizeHandle === 'se') {
-          // 대각선 리사이즈
-          newWidth = resizeStart.width + (wx - resizeStart.x);
-          newHeight = resizeStart.height + (wy - resizeStart.y);
-        } else if (resizeHandle === 'e') {
-          // 가로만 리사이즈
-          newWidth = resizeStart.width + (wx - resizeStart.x);
-        } else if (resizeHandle === 's') {
-          // 세로만 리사이즈
-          newHeight = resizeStart.height + (wy - resizeStart.y);
-        }
-
-        if (
-          newWidth > 10 &&
-          newHeight > 10 &&
-          newWidth < canvasSize.width * 2 &&
-          newHeight < canvasSize.height * 2
-        ) {
-          setImageSize({ width: newWidth, height: newHeight });
-          draw();
-        }
-        return;
-      }
-
-      // 이미지 드래그 중
-      if (isDraggingImage && !isImageFixed) {
-        const wx = (offsetX - viewPosRef.current.x) / scaleRef.current;
-        const wy = (offsetY - viewPosRef.current.y) / scaleRef.current;
-        setImagePosition({
-          x: wx - dragStart.x,
-          y: wy - dragStart.y,
-        });
-        draw();
-        return;
-      }
-
-      // 캔버스 팬닝 중
-      if (isPanningRef.current) {
-        viewPosRef.current = {
-          x: offsetX - startPosRef.current.x,
-          y: offsetY - startPosRef.current.y,
-        };
-        draw();
-      }
-      updateOverlay(offsetX, offsetY);
-    },
-    [
-      draw,
-      updateOverlay,
-      isDraggingImage,
-      isImageFixed,
-      dragStart,
-      isResizing,
-      resizeHandle,
-      resizeStart,
-      imageSize,
-      canvasSize,
-      setImagePosition,
-      setImageSize,
-    ]
-  );
-
-  const handleMouseUp = useCallback(
-    (e: React.MouseEvent<HTMLCanvasElement>) => {
-      // If it was a click (not a drag/pan)
-      if (dragStartInfoRef.current) {
-        const dx = e.nativeEvent.offsetX - dragStartInfoRef.current.x;
-        const dy = e.nativeEvent.offsetY - dragStartInfoRef.current.y;
-        if (Math.sqrt(dx * dx + dy * dy) <= DRAG_THRESHOLD) {
-          // This was a click, not a drag
-          const sx = e.nativeEvent.offsetX;
-          const sy = e.nativeEvent.offsetY;
-          const wx = (sx - viewPosRef.current.x) / scaleRef.current;
-          const wy = (sy - viewPosRef.current.y) / scaleRef.current;
-
-          const pixelX = Math.floor(wx);
-          const pixelY = Math.floor(wy);
-
-          if (
-            pixelX >= 0 &&
-            pixelX < canvasSize.width &&
-            pixelY >= 0 &&
-            pixelY < canvasSize.height &&
-            (!imageCanvasRef.current || isImageFixed) // Only allow pixel selection if no image or image is fixed
-          ) {
-            fixedPosRef.current = {
-              x: pixelX,
-              y: pixelY,
-              color: 'transparent',
-            };
-            setShowPalette(true);
-            centerOnPixel(sx, sy); // Call centerOnPixel here
-          }
-        }
-      }
-
-      isPanningRef.current = false;
-      setIsDraggingImage(false);
-      setIsResizing(false);
-      setResizeHandle(null);
-      dragStartInfoRef.current = null; // Reset drag start info
-    },
-    [canvasSize, isImageFixed, setShowPalette, centerOnPixel]
-  );
-
-  const handleMouseLeave = useCallback(
-    (e: React.MouseEvent<HTMLCanvasElement>) => {
-      handleMouseUp(e);
-      clearOverlay();
-      dragStartInfoRef.current = null; // Reset drag start info on mouse leave
-    },
-    [handleMouseUp, clearOverlay]
-  );
+  const {
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+    handleMouseLeave,
+    handleTouchStart,
+    handleTouchMove,
+    handleTouchEnd,
+  } = useCanvasInteraction({
+    viewPosRef,
+    scaleRef,
+    imageCanvasRef,
+    interactionCanvasRef,
+    fixedPosRef,
+    canvasSize,
+    imageMode,
+    isImageFixed,
+    isDraggingImage,
+    setIsDraggingImage,
+    dragStart,
+    setDragStart,
+    isResizing,
+    setIsResizing,
+    resizeHandle,
+    setResizeHandle,
+    resizeStart,
+    setResizeStart,
+    imagePosition,
+    setImagePosition,
+    imageSize,
+    setImageSize,
+    draw,
+    updateOverlay,
+    clearOverlay,
+    centerOnPixel,
+    getResizeHandle,
+    handleImageScale,
+    setShowPalette,
+    DRAG_THRESHOLD,
+    handleConfirm,
+  });
 
   // fetchCanvasData 분리
   useEffect(() => {
     fetchCanvasDataUtil({
       id: initialCanvasId,
       setIsLoading,
-      setHasError,
+      setHasError: (value: boolean) => {},
       setCanvasId,
       setCanvasSize,
       sourceCanvasRef,
@@ -895,142 +713,39 @@ function PixelCanvas({
     setCanvasId,
     setCanvasSize,
     setIsLoading,
-    setHasError,
     onLoadingChange,
     setShowCanvas,
   ]);
 
-  // 투명도 상태가 변경될 때 ref 값만 업데이트하고 draw 함수 직접 호출
-  const handleTransparencyChange = useCallback(
-    (value: number) => {
-      imageTransparencyRef.current = value;
-      setImageTransparency(value);
-      // 투명도가 변경되면 즉시 화면에 반영 (draw 함수 직접 호출)
-      if (imageCanvasRef.current) {
-        draw();
-      }
-    },
-    [draw, setImageTransparency]
-  );
+  // fetchCanvasData 분리
+  useEffect(() => {
+    fetchCanvasDataUtil({
+      id: initialCanvasId,
+      setIsLoading,
+      setHasError: (value: boolean) => {},
+      setCanvasId,
+      setCanvasSize,
+      sourceCanvasRef,
+      onLoadingChange,
+      setShowCanvas,
+      INITIAL_BACKGROUND_COLOR,
+    });
+  }, [
+    initialCanvasId,
+    setCanvasId,
+    setCanvasSize,
+    setIsLoading,
+    onLoadingChange,
+    setShowCanvas,
+  ]);
 
-  // PixelCanvas.tsx 내부에 아래 함수들을 추가합니다.
+  useEffect(() => {
+    if (initialCanvasId && initialCanvasId !== canvas_id) {
+      setCanvasId(initialCanvasId);
+      console.log('Canvas ID changed:', initialCanvasId);
+    }
+  }, [initialCanvasId, canvas_id, setCanvasId]);
 
-  // --- 터치 이벤트 핸들러 ---
-  const handleTouchStart = useCallback(
-    (e: React.TouchEvent<HTMLCanvasElement>) => {
-      e.preventDefault();
-      const touches = e.touches;
-
-      // 두 손가락 터치: 핀치 줌 시작
-      if (touches.length === 2) {
-        const dx = touches[0].clientX - touches[1].clientX;
-        const dy = touches[0].clientY - touches[1].clientY;
-        pinchDistanceRef.current = Math.sqrt(dx * dx + dy * dy);
-        isPanningRef.current = false; // 줌 할때는 패닝 방지
-        dragStartInfoRef.current = null; // 두 손가락 터치 시 드래그 시작 정보 초기화
-        return;
-      }
-
-      // 한 손가락 터치: 이동 또는 픽셀 선택 시작
-      if (touches.length === 1) {
-        const touch = touches[0];
-        const rect = interactionCanvasRef.current!.getBoundingClientRect();
-        const sx = touch.clientX - rect.left;
-        const sy = touch.clientY - rect.top;
-
-        dragStartInfoRef.current = { x: sx, y: sy };
-        lastTouchPosRef.current = { x: sx, y: sy };
-      }
-    },
-    []
-  );
-
-  const handleTouchMove = useCallback(
-    (e: React.TouchEvent<HTMLCanvasElement>) => {
-      e.preventDefault();
-      const touches = e.touches;
-      const rect = interactionCanvasRef.current!.getBoundingClientRect();
-
-      // 두 손가락 터치: 핀치 줌 로직
-      if (touches.length === 2) {
-        const dx = touches[0].clientX - touches[1].clientX;
-        const dy = touches[0].clientY - touches[1].clientY;
-        const newDistance = Math.sqrt(dx * dx + dy * dy);
-        const oldDistance = pinchDistanceRef.current;
-
-        if (oldDistance > 0) {
-          const scaleFactor = newDistance / oldDistance;
-          const newScale = Math.max(
-            MIN_SCALE,
-            Math.min(MAX_SCALE, scaleRef.current * scaleFactor)
-          );
-          const centerX =
-            (touches[0].clientX + touches[1].clientX) / 2 - rect.left;
-          const centerY =
-            (touches[0].clientY + touches[1].clientY) / 2 - rect.top;
-
-          const xs = (centerX - viewPosRef.current.x) / scaleRef.current;
-          const ys = (centerY - viewPosRef.current.y) / scaleRef.current;
-
-          viewPosRef.current.x = centerX - xs * newScale;
-          viewPosRef.current.y = centerY - ys * newScale;
-          scaleRef.current = newScale;
-
-          draw();
-          updateOverlay(centerX, centerY);
-        }
-        pinchDistanceRef.current = newDistance;
-        return;
-      }
-
-      // 한 손가락 터치: 이동 로직
-      if (touches.length === 1) {
-        const touch = touches[0];
-        const sx = touch.clientX - rect.left;
-        const sy = touch.clientY - rect.top;
-
-        if (dragStartInfoRef.current && !isPanningRef.current) {
-          const dx = sx - dragStartInfoRef.current.x;
-          const dy = sy - dragStartInfoRef.current.y;
-          if (Math.sqrt(dx * dx + dy * dy) > DRAG_THRESHOLD) {
-            isPanningRef.current = true;
-            startPosRef.current = {
-              x: sx - viewPosRef.current.x,
-              y: sy - viewPosRef.current.y,
-            };
-            dragStartInfoRef.current = null;
-          }
-        }
-
-        if (isPanningRef.current) {
-          viewPosRef.current = {
-            x: sx - startPosRef.current.x,
-            y: sy - startPosRef.current.y,
-          };
-          draw();
-        }
-        updateOverlay(sx, sy);
-        lastTouchPosRef.current = { x: sx, y: sy };
-      }
-    },
-    [draw, updateOverlay]
-  );
-
-  const handleTouchEnd = useCallback(
-    (e: React.TouchEvent<HTMLCanvasElement>) => {
-      // 모든 제스처 상태 초기화
-      pinchDistanceRef.current = 0;
-      // handleMouseUp에 마지막 터치 위치를 전달하여 클릭/드래그 판단에 사용
-      handleMouseUp({
-        nativeEvent: {
-          offsetX: lastTouchPosRef.current?.x || 0,
-          offsetY: lastTouchPosRef.current?.y || 0,
-        },
-      } as React.MouseEvent<HTMLCanvasElement>);
-      lastTouchPosRef.current = null; // 터치 종료 시 초기화
-    },
-    [handleMouseUp]
-  );
   // 투명도 상태가 변경될 때 ref 값 업데이트 및 draw 함수 호출
   useEffect(() => {
     imageTransparencyRef.current = imageTransparency;
@@ -1079,44 +794,6 @@ function PixelCanvas({
     observer.observe(rootElement);
     return () => observer.disconnect();
   }, [resetAndCenter]);
-
-  useEffect(() => {
-    const interactionCanvas = interactionCanvasRef.current;
-    if (!interactionCanvas) return;
-
-    const handleWheel = (e: WheelEvent) => {
-      e.preventDefault();
-      const { offsetX, offsetY } = e;
-
-      // 이미지 모드에서 이미지만 확대축소
-      if (imageMode && !isImageFixed && imageCanvasRef.current) {
-        const delta = -e.deltaY;
-        const scaleFactor = delta > 0 ? 1.1 : 0.9;
-        handleImageScale(scaleFactor);
-        return;
-      }
-
-      // 캔버스 모드 또는 이미지 확정 후 전체 확대축소
-      const xs = (offsetX - viewPosRef.current.x) / scaleRef.current;
-      const ys = (offsetY - viewPosRef.current.y) / scaleRef.current;
-      const delta = -e.deltaY;
-      const newScale =
-        delta > 0 ? scaleRef.current * 1.2 : scaleRef.current / 1.2;
-
-      if (newScale >= MIN_SCALE && newScale <= MAX_SCALE) {
-        scaleRef.current = newScale;
-        viewPosRef.current.x = offsetX - xs * scaleRef.current;
-        viewPosRef.current.y = offsetY - ys * scaleRef.current;
-        draw();
-        updateOverlay(offsetX, offsetY);
-      }
-    };
-
-    interactionCanvas.addEventListener('wheel', handleWheel, {
-      passive: false,
-    });
-    return () => interactionCanvas.removeEventListener('wheel', handleWheel);
-  }, [draw, updateOverlay, handleImageScale, imageMode, isImageFixed]);
 
   return (
     <div
